stop category occurrences from including child occurrences
This is potentially questionable behavior. It may be odd for parent categories to have a smaller number than the sum of its childrens'.

In the library table, we will still show all series' matching subcategories of the parent category selected, even if none of the series match the parent category.