improve handling of focus on reader page
Keyboard shortcuts will now no longer work when the page number text field is active, in order to avoid confusion when using them to edit the page number.

Additionally, the focus of the page is now set to the imageView (instead of the page number text field) which is what users would expect.